Basic operations into single file
