+++ test/unit/operations/operations.test.ts
@@ -1,4 +1,5 @@
-import { determineOperation } from '../../../src/lib/core/determineOperation'
+import { basicOperations, Operation } from '../../../src/lib/core/basicOperations'
+import { Blob } from '../../../src/lib/storage/Blob'
 import { TaskType, WacLdpTask } from '../../../src/lib/api/http/HttpParser'
 import { WacLdpResponse, ResultType } from '../../../src/lib/api/http/HttpResponder'
 import { toChunkStream } from '../helpers/toChunkStream'
@@ -12,7 +13,7 @@ test('delete blob', async () => {
     }),
     exists: () => true
   } as unknown as Blob
-  const operation = determineOperation(TaskType.blobDelete)
+  const operation: Operation = basicOperations(TaskType.blobDelete)
   const result: WacLdpResponse = await operation({} as WacLdpTask, node, false)
   expect((node as any).delete.mock.calls).toEqual([
     []
@@ -29,7 +30,7 @@ test.skip('write blob', async () => {
     }),
     exists: () => true
   } as unknown as Blob
-  const operation = determineOperation(TaskType.blobWrite)
+  const operation = basicOperations(TaskType.blobWrite)
   const result: WacLdpResponse = await operation({} as WacLdpTask, node, false)
   expect((node as any).setData.mock.calls).toEqual([
     []
@@ -46,7 +47,7 @@ test.skip('update blob', async () => {
     }),
     exists: () => true
   } as unknown as Blob
-  const operation = determineOperation(TaskType.blobUpdate)
+  const operation = basicOperations(TaskType.blobUpdate)
   const result: WacLdpResponse = await operation({} as WacLdpTask, node, false)
   expect((node as any).setData.mock.calls).toEqual([
     []
@@ -63,7 +64,7 @@ test('delete container', async () => {
     }),
     exists: () => true
   } as unknown as Container
-  const operation = determineOperation(TaskType.containerDelete)
+  const operation = basicOperations(TaskType.containerDelete)
   const result: WacLdpResponse = await operation({} as WacLdpTask, node, false)
   expect((node as any).delete.mock.calls).toEqual([
     []
@@ -80,7 +81,7 @@ test('read blob (omit body)', async () => {
     }),
     exists: () => true
   } as unknown as Blob
-  const operation = determineOperation(TaskType.blobRead)
+  const operation = basicOperations(TaskType.blobRead)
   const result: WacLdpResponse = await operation({ omitBody: true } as WacLdpTask, node, false)
   expect((node as any).getData.mock.calls).toEqual([
     []
@@ -102,7 +103,7 @@ test('read blob (with body)', async () => {
     }),
     exists: () => true
   } as unknown as Blob
-  const operation = determineOperation(TaskType.blobRead)
+  const operation = basicOperations(TaskType.blobRead)
   const result: WacLdpResponse = await operation({ omitBody: false } as WacLdpTask, node, false)
   expect((node as any).getData.mock.calls).toEqual([
     []
@@ -124,7 +125,7 @@ test('read container (omit body)', async () => {
     }),
     exists: () => true
   } as unknown as Container
-  const operation = determineOperation(TaskType.containerRead)
+  const operation = basicOperations(TaskType.containerRead)
   const result: WacLdpResponse = await operation({ path: '/foo', omitBody: true } as unknown as WacLdpTask, node, false)
   expect((node as any).getMembers.mock.calls).toEqual([
     []
@@ -147,7 +148,7 @@ test('read container (with body)', async () => {
     }),
     exists: () => true
   }as unknown as Container
-  const operation = determineOperation(TaskType.containerRead)
+  const operation = basicOperations(TaskType.containerRead)
   const result: WacLdpResponse = await operation({ path: '/foo', omitBody: false } as unknown as WacLdpTask, node, false)
   expect((node as any).getMembers.mock.calls).toEqual([
     []
